requirements.txt
flask-mysql
-U flask-cors
-U Flask-SQLAlchemy
flask-bcrypt
